Feat #4: Add docker support
It will expose only `code` folder to docker, while target will be inside the image.

Idea is that developers should be able to build images with installed packages inside and then just use it without having to download all of the packages again.. when new package is added, we can add it to dummy app if we want to avoid having to download it after destroying of a container.